+++ src/main/java/com/aps/projectmanage/domain/repository/ProjectRepository.java
@@ -4,5 +4,5 @@ import com.aps.projectmanage.domain.entity.Project;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProjectRepository extends BaseRepository<Project, Long>{
+public interface ProjectRepository extends BaseRepository<Project, Integer>{
 }
