+++ src/main/java/com/aps/projectmanage/domain/repository/TaskCommentRepository.java
@@ -4,5 +4,5 @@ import com.aps.projectmanage.domain.entity.TaskComment;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TaskCommentRepository extends BaseRepository<TaskComment, Long> {
+public interface TaskCommentRepository extends BaseRepository<TaskComment, Integer> {
 }
