+++ src/main/java/com/aps/projectmanage/domain/repository/UserRepository.java
@@ -4,5 +4,5 @@ import com.aps.projectmanage.domain.entity.User;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends BaseRepository<User, Long>  {
+public interface UserRepository extends BaseRepository<User, Integer>  {
 }
